Move existing tests to corresponding files
